+++ core/src/main/java/com/expediagroup/streamplatform/streamregistry/core/services/DomainService.java
@@ -30,6 +30,7 @@ import com.expediagroup.streamplatform.streamregistry.core.handlers.HandlerServi
 import com.expediagroup.streamplatform.streamregistry.core.validators.DomainValidator;
 import com.expediagroup.streamplatform.streamregistry.core.validators.ValidationException;
 import com.expediagroup.streamplatform.streamregistry.model.Domain;
+import com.expediagroup.streamplatform.streamregistry.model.Status;
 import com.expediagroup.streamplatform.streamregistry.model.keys.DomainKey;
 import com.expediagroup.streamplatform.streamregistry.repository.DomainRepository;
 
@@ -40,7 +41,7 @@ public class DomainService {
   private final DomainValidator domainValidator;
   private final DomainRepository domainRepository;
 
-  @PreAuthorize("hasPermission(#domain, 'create')")
+  @PreAuthorize("hasPermission(#domain, 'CREATE')")
   public Optional<Domain> create(Domain domain) throws ValidationException {
     if (read(domain.getKey()).isPresent()) {
       throw new ValidationException("Can't create because it already exists");
@@ -50,7 +51,7 @@ public class DomainService {
     return save(domain);
   }
 
-  @PreAuthorize("hasPermission(#domain, 'update')")
+  @PreAuthorize("hasPermission(#domain, 'UPDATE')")
   public Optional<Domain> update(Domain domain) throws ValidationException {
     var existing = read(domain.getKey());
     if (!existing.isPresent()) {
@@ -58,6 +59,13 @@ public class DomainService {
     }
     domainValidator.validateForUpdate(domain, existing.get());
     domain.setSpecification(handlerService.handleUpdate(domain, existing.get()));
+    return save(domain);
+  }
+
+  @PreAuthorize("hasPermission(#domainKey, 'UPDATE_STATUS')")
+  public Optional<Domain> updateStatus(DomainKey domainKey, Status status) {
+    Domain domain = read(domainKey).get();
+    domain.setStatus(status);
     return save(domain);
   }
 
@@ -78,7 +86,7 @@ public class DomainService {
     return domainRepository.findAll().stream().filter(filter).collect(toList());
   }
 
-  @PreAuthorize("hasPermission(#domain, 'delete')")
+  @PreAuthorize("hasPermission(#domain, 'DELETE')")
   public void delete(Domain domain) {
     throw new UnsupportedOperationException();
   }