+++ core/src/main/java/com/expediagroup/streamplatform/streamregistry/core/services/SchemaService.java
@@ -30,6 +30,7 @@ import com.expediagroup.streamplatform.streamregistry.core.handlers.HandlerServi
 import com.expediagroup.streamplatform.streamregistry.core.validators.SchemaValidator;
 import com.expediagroup.streamplatform.streamregistry.core.validators.ValidationException;
 import com.expediagroup.streamplatform.streamregistry.model.Schema;
+import com.expediagroup.streamplatform.streamregistry.model.Status;
 import com.expediagroup.streamplatform.streamregistry.model.keys.SchemaKey;
 import com.expediagroup.streamplatform.streamregistry.repository.SchemaRepository;
 
@@ -40,7 +41,7 @@ public class SchemaService {
   private final SchemaValidator schemaValidator;
   private final SchemaRepository schemaRepository;
 
-  @PreAuthorize("hasPermission(#schema, 'create')")
+  @PreAuthorize("hasPermission(#schema, 'CREATE')")
   public Optional<Schema> create(Schema schema) throws ValidationException {
     if (read(schema.getKey()).isPresent()) {
       throw new ValidationException("Can't create because it already exists");
@@ -50,7 +51,7 @@ public class SchemaService {
     return save(schema);
   }
 
-  @PreAuthorize("hasPermission(#schema, 'update')")
+  @PreAuthorize("hasPermission(#schema, 'UPDATE')")
   public Optional<Schema> update(Schema schema) throws ValidationException {
     var existing = read(schema.getKey());
     if (!existing.isPresent()) {
@@ -58,6 +59,13 @@ public class SchemaService {
     }
     schemaValidator.validateForUpdate(schema, existing.get());
     schema.setSpecification(handlerService.handleUpdate(schema, existing.get()));
+    return save(schema);
+  }
+
+  @PreAuthorize("hasPermission(#schemaKey, 'UPDATE_STATUS')")
+  public Optional<Schema> updateStatus(SchemaKey schemaKey, Status status) {
+    Schema schema = read(schemaKey).get();
+    schema.setStatus(status);
     return save(schema);
   }
 
@@ -78,7 +86,7 @@ public class SchemaService {
     return schemaRepository.findAll().stream().filter(filter).collect(toList());
   }
 
-  @PreAuthorize("hasPermission(#schema, 'delete')")
+  @PreAuthorize("hasPermission(#schema, 'DELETE')")
   public void delete(Schema schema) {
     throw new UnsupportedOperationException();
   }